Adjust: Shield controller + shield
I changed how the shield works / is used. It has health and the plan is to break it and stun the player. As of now enemies cannot hit player, but needs to lower the health of the shield. Since there is only one shield that is unique to the player, make a singleton so other enemies can access it and lower its health.